requirements.txt
aioquic==1.2.0
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
cryptography==45.0.2
h11==0.16.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pylsqpack==0.3.22
pyOpenSSL==25.1.0
PySocks==1.7.1
scapy==2.6.1
service-identity==24.2.0